microservices/requirements.txt
fastapi==0.115.12
langserve==0.3.1
numpy==2.2.4
pydantic==2.11.2
python-dotenv==1.1.0
uvicorn==0.34.0
